[DR-3034] Delete policies in snapshot delete flight (#1459)
* Delete policy from TPS when deleting snapshot

* Add unit tests for DeleteSnapshotPolicyStep

* Test TPS not enabled when creating policies

For consistency, also add a test for the
CreateSnapshotPolicyStep that verifies the step
will not fail if the policy service is not yet
enabled.

* Only call deletePao for snapshots using secure monitoring

* Fix deleting snapshots that do not exist

If a snapshot does not exist, we assume that
its policy was already deleted along with the
snapshot. By wrapping the step with PerformDatasetStep,
we only call DeleteSnapshotPolicyStep if the source
dataset exists (since we need to check that the source
dataset has secure monitoring enabled).

* Handle FeatureNotImplementedException in CreateSnapshotPolicyStep undo

* Re-create policy if snapshot delete fails

* Rename deletePao and add docstring

* Refactor feature-enabled check

Instead of throwing a FeatureNotImplementedException from
the PolicyService methods and catching the exception where
they get called, have the methods return early if TPS is
not enabled.

* Consolidate snapshot policy create/delete steps

* Fix typo

+++ src/main/java/bio/terra/service/policy/PolicyService.java
@@ -50,7 +50,10 @@ public class PolicyService {
 
   public void createPao(
       UUID resourceId, TpsObjectType resourceType, @Nullable TpsPolicyInputs policyInputs) {
-    policyServiceConfiguration.tpsEnabledCheck();
+    if (!policyServiceConfiguration.getEnabled()) {
+      logger.info("Terra Policy Service is not enabled.");
+      return;
+    }
     TpsPolicyInputs inputs = (policyInputs == null) ? new TpsPolicyInputs() : policyInputs;
 
     TpsApi tpsApi = policyApiService.getPolicyApi();
@@ -66,13 +69,19 @@ public class PolicyService {
     }
   }
 
-  public void deletePao(UUID resourceId) {
-    policyServiceConfiguration.tpsEnabledCheck();
+  /**
+   * Delete the policy access object by its id. If it does not exist in the policy service, ignore
+   * the PolicyServiceNotFoundException exception.
+   */
+  public void deletePaoIfExists(UUID resourceId) {
+    if (!policyServiceConfiguration.getEnabled()) {
+      logger.info("Terra Policy Service is not enabled.");
+      return;
+    }
     TpsApi tpsApi = policyApiService.getPolicyApi();
     try {
       tpsApi.deletePao(resourceId);
     } catch (ApiException e) {
-      // Ignore the exception if the policy object being deleted does not exist
       RuntimeException exception = convertApiException(e);
       if (!(exception instanceof PolicyServiceNotFoundException)) {
         throw exception;
